requirements_condalist.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: linux-64
_libgcc_mutex=0.1=main
_openmp_mutex=5.1=1_gnu
biopython=1.78=py36h7f8727e_0
blas=1.0=mkl
ca-certificates=2023.7.22=hbcca054_0
cached-property=1.5.2=pypi_0
captum=0.6.0=0
certifi=2016.9.26=py36_0
colorama=0.4.4=pyhd3eb1b0_0
cudatoolkit=11.1.1=ha002fc5_10
cycler=0.11.0=pyhd3eb1b0_0
dataclasses=0.8=pyh787bdff_2
dbus=1.13.18=hb2f20db_0
einops=0.4.1=pypi_0
expat=2.4.4=h295c915_0
fontconfig=2.13.1=h6c09931_0
freetype=2.11.0=h70c0345_0
glib=2.69.1=h4ff587b_1
gst-plugins-base=1.14.0=h8213a91_2
gstreamer=1.14.0=h28cd5cc_2
h5py=3.1.0=pypi_0
icu=58.2=he6710b0_3
igraph=0.9.11=pypi_0
importlib-resources=5.4.0=pypi_0
intel-openmp=2022.0.1=h06a4308_3633
joblib=1.0.1=pyhd3eb1b0_0
jpeg=9e=h7f8727e_0
kiwisolver=1.3.1=py36h2531618_0
ld_impl_linux-64=2.38=h1181459_1
leidenalg=0.8.10=pypi_0
libffi=3.3=he6710b0_2
libgcc-ng=11.2.0=h1234567_1
libgfortran-ng=7.5.0=ha8ba4b0_17
libgfortran4=7.5.0=ha8ba4b0_17
libgomp=11.2.0=h1234567_1
libpng=1.6.37=hbc83047_0
libstdcxx-ng=11.2.0=h1234567_1
libuuid=1.0.3=h7f8727e_2
libuv=1.44.2=h5eee18b_0
libxcb=1.15=h7f8727e_0
libxml2=2.9.14=h74e7548_0
matplotlib=3.3.4=pypi_0
mkl=2020.2=256
mkl-service=2.3.0=py36he8ac12f_0
mkl_fft=1.3.0=py36h54f3939_0
mkl_random=1.1.1=py36h0573a6f_0
modisco=0.5.16.0=pypi_0
ncurses=6.3=h7f8727e_2
ninja=1.7.2=0
numpy=1.19.5=pypi_0
numpy-base=1.19.2=py36hfa32c7d_0
openssl=1.1.1l=h7f98852_0
pandas=1.1.5=py36ha9443f7_0
pcre=8.45=h295c915_0
pillow=8.4.0=pypi_0
pip=21.2.2=py36h06a4308_0
psutil=5.9.5=pypi_0
pyparsing=3.0.4=pyhd3eb1b0_0
pyqt=5.9.2=py36h05f1152_2
python=3.6.13=h12debd9_1
python-dateutil=2.8.2=pyhd3eb1b0_0
pytorch=1.8.1=py3.6_cuda11.1_cudnn8.0.5_0
pytz=2021.3=pyhd3eb1b0_0
qt=5.9.7=h5867ecd_1
readline=8.1.2=h7f8727e_1
regex=2023.8.8=pypi_0
scikit-learn=0.24.2=py36ha9443f7_0
scipy=1.5.2=py36h0b6359f_0
seaborn=0.11.2=pypi_0
setuptools=58.0.4=py36h06a4308_0
sip=4.19.8=py36hf484d3e_0
six=1.16.0=pyhd3eb1b0_1
sqlite=3.38.5=hc218d9a_0
texttable=1.6.7=pypi_0
threadpoolctl=2.2.0=pyh0d69192_0
tk=8.6.12=h1ccaba5_0
torch=1.10.2+cu113=pypi_0
torchaudio=0.10.2+cu113=pypi_0
torchvision=0.11.3+cu113=pypi_0
tornado=6.1=py36h27cfd23_0
tqdm=4.63.0=pyhd3eb1b0_0
typing-extensions=4.1.1=pypi_0
typing_extensions=3.10.0.2=pyha770c72_0
wheel=0.37.1=pyhd3eb1b0_0
xlrd=1.1.0=py36_0
xz=5.2.5=h7f8727e_1
zipp=3.6.0=pypi_0
zlib=1.2.12=h7f8727e_2
